# specs/001-hybridpeaks-mvp/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: HybridPeaks API
  description: REST API for HybridPeaks training platform MVP
  version: 1.0.0
  contact:
    name: HybridPeaks Team
servers:
  - url: https://api.hybridpeaks.com/v1
    description: Production server
  - url: https://api-staging.hybridpeaks.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Create a new user account (coach or athlete)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: Authenticate user and receive tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchange refresh token for new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Retrieve authenticated user profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== PLANNING (COACHES) ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, role, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
          description: Min 12 chars, must include uppercase, lowercase, digit
        role:
          type: string
          enum: [coach, athlete]
        name:
          type: string
          minLength: 2
          maxLength: 100
        coachId:
          type: string
          format: uuid
          description: Required if role is athlete
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (15 min expiry)
        refreshToken:
          type: string
          description: Refresh token (7 day expiry)
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum: [coach, athlete]
        name:
          type: string
        coachId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ==================== TRAINING PLAN SCHEMAS ====================
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...